# PERF: use groupby.transform fast path for DataFrame -> Series aggregations